requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.2.0
Augmentor==0.2.10
ax-platform==0.2.5.1
botorch==0.6.4
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cryptography==38.0.4
cycler==0.11.0
Cython==0.29.33
distlib==0.3.6
editdistance==0.6.2
filelock==3.10.0
Flask==2.2.3
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
fvcore==0.1.5.post20220512
gevent==22.10.2
gevent-websocket==0.10.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
gpytorch==1.8.1
greenlet==2.0.2
grpcio==1.43.0
hydra-core==1.2.0
idna==2.10
imageio==2.26.0
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jpeg2dct==0.2.4
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.1
lmdb==1.3.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
msgpack==1.0.5
multidict==6.0.4
multipledispatch==0.6.0
networkx==3.0
nltk==3.7
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.2.0
platformdirs==3.1.1
plotly==5.13.1
pluggy==1.0.0
portalocker==2.7.0
prettytable==3.6.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==2.1.2
pyDeprecate==0.3.2
pyOpenSSL==19.1.0
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
PyTurboJPEG==1.7.0
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
setuptools==59.5.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
shapely==2.0.1
six==1.16.0
sortedcontainers==2.4.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.2.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.3.15
timm==0.6.5
toolz==0.12.0
torchmetrics==0.11.4
tqdm==4.64.0
typeguard==3.0.1
typing_extensions==4.5.0
urllib3==1.25.11
virtualenv==20.21.0
wcwidth==0.2.6
Werkzeug==2.2.3
yacs==0.1.8
yapf==0.32.0
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
zstandard==0.18.0